Renaming with more official terminology
