+++ demo/pom.xml
@@ -2,8 +2,8 @@
 <project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>org.acme</groupId>
-  <artifactId>getting-started</artifactId>
+  <groupId>com.redhat.swatch</groupId>
+  <artifactId>billing-provider-aws</artifactId>
   <version>1.0.0-SNAPSHOT</version>
   <properties>
     <clowder-quarkus-config-source.version>0.5.3</clowder-quarkus-config-source.version>
