+++ demo/src/main/resources/application.properties
@@ -3,10 +3,7 @@ DB_PORT=${clowder.database.port:5432}
 DB_NAME=${clowder.database.name:rhsm-subscriptions}
 DB_USER=${clowder.database.username:rhsm-subscriptions}
 DB_PASSWORD=${clowder.database.password:rhsm-subscriptions}
-SERVER_PORT=${clowder.endpoints.lburnettquarkus.port:8000}
-KAFKA_BROKER_HOST=${clowder.kafka.brokers[0].hostname:localhost}
-KAFKA_BROKER_PORT=${clowder.kafka.brokers[0].port:9092}
-KAFKA_TOPIC_NAME=${clowder.kafka.topics[0].name:fruit-salad-requests}
+SERVER_PORT=${clowder.endpoints.swatch-billing-provider-aws.port:8000}
 
 quarkus.http.port=${SERVER_PORT}
 
@@ -20,10 +17,16 @@ quarkus.liquibase.migrate-at-start=true
 quarkus.liquibase.change-log=db/changelog/db.changelog-master.yaml
 quarkus.liquibase.clean-at-start=true
 
-kafka.bootstrap.servers=${KAFKA_BROKER_HOST}:${KAFKA_BROKER_PORT}
-# Set the Kafka topic, as it's not the channel name. "requests" needs to match the @Incoming tag on the processor
-mp.messaging.incoming.requests.topic=${KAFKA_TOPIC_NAME}
 
+#clowder quarkus config takes care of setting these, no need to try to do clowder.kafka.brokers[0]
 
-# Go bad to the first records, if it's out first access
+kafka.bootstrap.servers=localhost:9092
+
+mp.messaging.incoming.ingress.connector=smallrye-kafka
+mp.messaging.incoming.ingress.topic=platform.rhsm-subscriptions.tally
+
+mp.messaging.outgoing.egress.connector=smallrye-kafka
+mp.messaging.outgoing.egress.topic=platform.rhsm-subscriptions.tally
+
+# Go back to the first records, if it's out first access
 kafka.auto.offset.reset=earliest
