+++ demo/src/test/java/com/redhat/swatch/FruitResourceTest.java
@@ -1,4 +1,4 @@
-package org.acme;
+package com.redhat.swatch;
 
 import io.quarkus.test.junit.QuarkusTest;
 import org.junit.jupiter.api.Test;
